start route, claim badges, finish/interupt journey and give a rating

+++ database/migrations/2024_05_22_121637_create_rates_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('rates', function (Blueprint $table) {
             $table->id();
-            $table->integer('rate');
+            $table->float('rate');
             $table->foreignId('route_id')->constrained()->onDelete('cascade');
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->timestamps();
